major changes to switch from SAM to BAM
