+++ .github/workflows/gottcha.yml
@@ -51,29 +51,35 @@ jobs:
           cd GOTTCHA2/test/
           gottcha2 profile -d gottcha2_database_test/gottcha_db.species.fna -i SRR12689945_1_1k.fastq.gz -nc -o output
           gottcha2 profile -d gottcha2_database_test/gottcha_db.species.fna -i gottcha2_database_test/gottcha_db.species.fna.gz -np -o output
-          gottcha2 profile -d gottcha2_database_test/gottcha_db.species.fna -s output/SRR12689945_1_1k.gottcha_species.sam -o output
+          gottcha2 profile -d gottcha2_database_test/gottcha_db.species.fna -b output/SRR12689945_1_1k.gottcha_species.bam -o output
+          grep 'Orthoebolavirus zairense' output/SRR12689945_1_1k.tsv
+
+      - name: Run testing dataset - Profile tests with additional options
+        run: |
+          cd GOTTCHA2/test/
+          gottcha2 profile -d gottcha2_database_test/gottcha_db.species.fna -i SRR12689945_1_1k.fastq.gz -mi 0.9 -mf 0.9 -mg 100 -o output
+          gottcha2 profile -d gottcha2_database_test/gottcha_db.species.fna -b output/SRR12689945_1_1k.gottcha_species.bam -a gottcha2_database_test/aoi_list.txt -o output
           grep 'Orthoebolavirus zairense' output/SRR12689945_1_1k.tsv
 
       - name: Run testing dataset - ONT profile tests
         run: |
           cd GOTTCHA2/test/
           gottcha2 profile -d gottcha2_database_test/gottcha_db.species.fna -i SRR2016557_extract.fasta.gz -o output -np -er 0.05
-          gottcha2.py -d gottcha2_database_test/gottcha_db.species.fna -i SRR2016557_extract.fasta.gz -o output -np -er 0.05
           grep 'Orthoebolavirus zairense' output/SRR2016557_extract.tsv
 
       - name: Test extract functionality - Simple extraction
         run: |
           cd GOTTCHA2/test/
-          gottcha2 extract -d gottcha2_database_test/gottcha_db.species.fna -e 3052462 -s output/SRR12689945_1_1k.gottcha_species.sam -o output
+          gottcha2 extract -d gottcha2_database_test/gottcha_db.species.fna -e 3052462 -b output/SRR12689945_1_1k.gottcha_species.bam -o output
           count_tsv=$(grep "3052462" output/SRR12689945_1_1k.tsv | cut -f4)
           grep -c '>' output/SRR12689945_1_1k.gottcha_species.extract.fasta | grep $count_tsv
 
       - name: Test extract functionality - Additional options
         run: |
           cd GOTTCHA2/test/
-          gottcha2 extract -d gottcha2_database_test/gottcha_db.species.fna -e 3052462:10:fastq -s output/SRR12689945_1_1k.gottcha_species.sam -p test_ext1 -o output
+          gottcha2 extract -d gottcha2_database_test/gottcha_db.species.fna -e 3052462:10:fastq -b output/SRR12689945_1_1k.gottcha_species.bam -p test_ext1 -o output
           wc -l output/test_ext1.extract.fastq | grep 40
-          gottcha2 extract -d gottcha2_database_test/gottcha_db.species.fna -ef -eo -s output/SRR12689945_1_1k.gottcha_species.sam -p test_ext2 -o output
+          gottcha2 extract -d gottcha2_database_test/gottcha_db.species.fna -ef -eo -b output/SRR12689945_1_1k.gottcha_species.bam -p test_ext2 -o output
           wc -l output/test_ext2.extract.fasta | grep 320
 
       - name: Run unittests
