+++ gottcha/__init__.py
@@ -1,3 +1,1 @@
-__version__ = "2.2.2"
-
 GOTTCHA_DB_LATEST = "https://ref-db.edgebioinformatics.org/gottcha2/latest/gottcha_db.species.tar"
