--- test/test_gottcha2_integration.py
@@ -1,235 +1,0 @@
-import unittest
-import os
-import sys
-import tempfile
-import shutil
-import pandas as pd
-from unittest.mock import patch, MagicMock
-from pathlib import Path
-import time
-
-# Get the path to the project root directory
-current_dir = Path(__file__).parent
-project_root = current_dir.parent
-sys.path.insert(0, str(project_root))
-
-# Import the module
-import gottcha.utils.gottcha2 as gottcha2
-
-class TestGottcha2Integration(unittest.TestCase):
-    """Integration tests for GOTTCHA2."""
-    
-    def setUp(self):
-        """Set up test environment."""
-        self.test_dir = tempfile.mkdtemp()
-        self.test_data_dir = os.path.join(os.path.dirname(os.path.abspath(__file__)), 'test_data')
-        
-        # Create a mock SAM file with proper SAM format
-        self.sam_file = os.path.join(self.test_dir, "test.sam")
-        with open(self.sam_file, 'w') as f:
-            # Header line (required for SAM format)
-            f.write("@HD\tVN:1.0\n")
-            # Sequence dictionary lines
-            f.write("@SQ\tSN:ABC|1|100|12345\tLN:100\n")
-            f.write("@SQ\tSN:XYZ|1|100|67890\tLN:100\n")
-            # Actual alignment records
-            f.write("read1\t0\tABC|1|100|12345\t11\t60\t5S10M3S\t*\t0\t0\tGGGGGCCCCCCCCCGGG\tHHHHHHHHHHHHHHHHH\tNM:i:0\tAS:i:10\tXS:i:0\n")
-            f.write("read2\t0\tABC|1|100|12345\t21\t60\t5S10M3S\t*\t0\t0\tGGGGGCCCCCCCCCGGG\tHHHHHHHHHHHHHHHHH\tNM:i:0\tAS:i:10\tXS:i:0\n")
-            f.write("read3\t0\tXYZ|1|100|67890\t31\t60\t5S10M3S\t*\t0\t0\tGGGGGCCCCCCCCCGGG\tHHHHHHHHHHHHHHHHH\tNM:i:0\tAS:i:10\tXS:i:0\n")
-    
-    def tearDown(self):
-        """Clean up after tests."""
-        shutil.rmtree(self.test_dir)
-    
-    @patch('gottcha.scripts.gottcha2.Pool')
-    def test_process_sam_file(self, mock_pool):
-        """Test processing of SAM file."""
-        # Mock the Pool and its return values
-        mock_job = MagicMock()
-        mock_pool.return_value.apply_async.return_value = mock_job
-        
-        # Set up mock data to return from worker
-        mock_result = ({
-            'ABC|1|100|12345': {
-                'REGIONS': [(11, 20), (21, 30)],
-                'MB': 20,
-                'MR': 2,
-                'NM': 0,
-                'ID': 0,
-                'RL': 34
-            },
-            'XYZ|1|100|67890': {
-                'REGIONS': [(31, 40)],
-                'MB': 10,
-                'MR': 1,
-                'NM': 0,
-                'ID': 0,
-                'RL': 17
-            }
-        }, 3, 0)  # Return tuple matching current worker signature
-        
-        mock_job.get.return_value = mock_result
-        
-        # Create an actual test file with gottcha2 global variables
-        gottcha2.argvs = MagicMock()
-        gottcha2.argvs.debug = False
-        gottcha2.argvs.silent = True
-        gottcha2.begin_t = 0
-        gottcha2.logfile = os.path.join(self.test_dir, "test.log")
-        
-        # Run the function
-        result, mapped_reads, tol_alignment_count, tol_invalid_match_count = gottcha2.process_sam_file(self.sam_file, 1, 0.5, 0, 0)
-        
-        # Check the results
-        self.assertEqual(mapped_reads, 3)  # 2 + 1 from the mocked data
-        self.assertEqual(len(result), 2)   # Two reference sequences
-        self.assertEqual(result['ABC|1|100|12345']['SC'], 20)  # Sum of coverage for merged regions
-        self.assertEqual(result['XYZ|1|100|67890']['SC'], 10)
-        self.assertEqual(tol_alignment_count, 3)
-        self.assertEqual(tol_invalid_match_count, 0)
-
-    def test_group_refs_to_strains(self):
-        """Test grouping references by strain."""
-        # Set up mock database stats with required columns/index
-        gottcha2.df_stats = pd.DataFrame([
-            ['species', '12345', 1000, 10000, ''],
-            ['species', '67890', 500, 5000, '']
-        ], columns=['DB_level', 'Taxid', 'TotalLength', 'GenomeSize', 'Note'])
-        gottcha2.df_stats.set_index('Taxid', inplace=True)
-        
-        # Test the function with mock data
-        test_data = {
-            'ABC|1|100|12345': {'MB': 100, 'MR': 10, 'NM': 2, 'ID': 1, 'SC': 50, 'RL': 200},
-            'XYZ|1|100|67890': {'MB': 50, 'MR': 5, 'NM': 1, 'ID': 0, 'SC': 25, 'RL': 100}
-        }
-        
-        str_df, aoi_read_count = gottcha2.group_refs_to_strains(test_data, None, 'report')
-        self.assertEqual(len(str_df), 2)
-        self.assertEqual(aoi_read_count, 0)
-        self.assertTrue({'TAXID', 'TOTAL_BP_MAPPED', 'TOTAL_BP_INDEL', 'AOI_READ_COUNT'}.issubset(str_df.columns))
-
-    def test_taxid_file_input(self):
-        """Test extracting reads using taxids from a file."""
-        # Create a temporary taxid file
-        taxid_file = os.path.join(self.test_dir, "taxids.txt")
-        with open(taxid_file, 'w') as f:
-            f.write("12345\n")
-            f.write("67890\n")
-            f.write("\n")  # Empty line to test handling
-            f.write("# Comment to ignore\n")
-        
-        # Create mock taxonomy file data
-        mock_taxa_df = pd.DataFrame({
-            'LEVEL': ['species', 'species'],
-            'NAME': ['Species1', 'Species2'],
-            'TAXID': ['12345', '67890'],
-            'NOTE': ['', '']
-        })
-
-        # Mock logging and file operations
-        with patch('gottcha.scripts.gottcha2.print_message') as mock_print, \
-             patch('gottcha.scripts.gottcha2.pd.read_csv', return_value=mock_taxa_df):
-            
-            # Test with empty DataFrame and file input
-            gottcha2.argvs = MagicMock()
-            gottcha2.argvs.silent = True
-            gottcha2.begin_t = 0
-            gottcha2.logfile = os.path.join(self.test_dir, "test.log")
-            
-            empty_df = pd.DataFrame()
-            taxa_dict, qualified_taxids = gottcha2.parse_taxids(f'@{taxid_file}', empty_df, "mock_full.tsv")
-            
-            # Verify the results
-            self.assertEqual(len(qualified_taxids), 2)
-            self.assertListEqual(sorted(qualified_taxids), ['12345', '67890'])
-            self.assertEqual(len(taxa_dict), 2)
-            
-            # Verify taxonomy dictionary contents
-            self.assertEqual(taxa_dict['12345']['level'], 'species')
-            self.assertEqual(taxa_dict['12345']['name'], 'Species1')
-            self.assertEqual(taxa_dict['67890']['level'], 'species')
-            self.assertEqual(taxa_dict['67890']['name'], 'Species2')
-
-            # Test with pre-loaded DataFrame
-            taxa_dict2, qualified_taxids2 = gottcha2.parse_taxids(f'@{taxid_file}', mock_taxa_df, "mock_full.tsv")
-            self.assertEqual(len(qualified_taxids2), 2)
-            self.assertListEqual(sorted(qualified_taxids2), ['12345', '67890'])
-            
-            # Test with direct taxid input
-            taxa_dict3, qualified_taxids3 = gottcha2.parse_taxids('12345,67890', mock_taxa_df, "mock_full.tsv")
-            self.assertEqual(len(qualified_taxids3), 2)
-            self.assertListEqual(sorted(qualified_taxids3), ['12345', '67890'])
-            
-            # Test with 'all' option
-            taxa_dict4, qualified_taxids4 = gottcha2.parse_taxids('all', mock_taxa_df, "mock_full.tsv")
-            self.assertEqual(len(qualified_taxids4), 2)
-            self.assertListEqual(sorted(qualified_taxids4), ['12345', '67890'])
-
-    def test_load_acc_list(self):
-        """Test loading accession list to exclude."""
-        # Create a temporary file with accessions
-        acc_file = os.path.join(self.test_dir, "exclude_acc.txt")
-        with open(acc_file, 'w') as f:
-            f.write("ABC\nXYZ\nPQR\n")
-        
-        # Test loading the file
-        exclude_set = gottcha2.load_acc_list(acc_file)
-        
-        self.assertEqual(len(exclude_set), 3)
-        self.assertIn("ABC", exclude_set)
-        self.assertIn("XYZ", exclude_set)
-        self.assertIn("PQR", exclude_set)
-        
-        # Test loading empty file
-        empty_file = os.path.join(self.test_dir, "empty.txt")
-        with open(empty_file, 'w') as f:
-            pass
-            
-        with self.assertLogs(level='WARNING'):
-            empty_set = gottcha2.load_acc_list(empty_file)
-                
-        self.assertEqual(len(empty_set), 0)
-
-    def test_exclude_accessions_in_parse(self):
-        """Test parse() output fields and validity flags."""
-        # Create SAM line with accession ABC
-        sam_line = "read1\t0\tABC|1|100|12345\t11\t60\t5S10M3S\t*\t0\t0\tGGGGGCCCCCCCCCGGG\tHHHHHHHHHHHHHHHHH\tNM:i:0\tAS:i:10\tXS:i:0\tZC:i:1"
-        
-        # Current parse signature: (line, matchFraction, matchIdentity, matchLength)
-        _, region1, _, _, _, _, _, _, _, _, _, valid_match_flag1, zc_flag = gottcha2.parse(sam_line, 0.5, 0.0, 0)
-        self.assertTrue(valid_match_flag1)
-        self.assertTrue(zc_flag)
-        self.assertEqual(region1, (11, 20))
-        
-        # Higher required match fraction should invalidate this alignment
-        _, _, _, _, _, _, _, _, _, _, _, valid_match_flag2, zc_flag2 = gottcha2.parse(sam_line, 0.95, 0.0, 0)
-        self.assertFalse(valid_match_flag2)
-        self.assertTrue(zc_flag2)
-        
-        # Test with a different accession and no split-read chunk tag
-        sam_line2 = "read3\t0\tDEF|1|100|67890\t11\t60\t5S10M3S\t*\t0\t0\tGGGGGCCCCCCCCCGGG\tHHHHHHHHHHHHHHHHH\tNM:i:0\tAS:i:10\tXS:i:0"
-        _, region3, _, _, _, _, _, _, _, _, _, valid_match_flag3, zc_flag3 = gottcha2.parse(sam_line2, 0.5, 0.0, 0)
-        self.assertEqual(region3, (11, 20))
-        self.assertTrue(valid_match_flag3)
-        self.assertFalse(zc_flag3)
-
-    def test_worker_with_excluded_acc_list(self):
-        """Test that worker() correctly excludes accessions."""
-        # Create a temporary SAM file with multiple accessions
-        test_sam = os.path.join(self.test_dir, "acc_test.sam")
-        with open(test_sam, 'w') as f:
-            f.write("read1\t0\tABC|1|100|12345\t11\t60\t5S10M3S\t*\t0\t0\tGGGGGCCCCCCCCCGGG\tHHHHHHHHHHHHHHHHH\tNM:i:0\tAS:i:10\tXS:i:0\n")
-            f.write("read2\t0\tXYZ|1|100|67890\t21\t60\t5S10M3S\t*\t0\t0\tGGGGGCCCCCCCCCGGG\tHHHHHHHHHHHHHHHHH\tNM:i:0\tAS:i:10\tXS:i:0\n")
-            f.write("read3\t0\tDEF|1|100|54321\t31\t60\t5S10M3S\t*\t0\t0\tGGGGGCCCCCCCCCGGG\tHHHHHHHHHHHHHHHHH\tNM:i:0\tAS:i:10\tXS:i:0\n")
-
-        # Test without exclusion
-        result_no_exclude, lines_count1, invalid_match_count1 = gottcha2.worker(test_sam, 0, os.path.getsize(test_sam), 0.5, 0, 0, False)
-        self.assertEqual(len(result_no_exclude), 3)
-        self.assertIn("ABC|1|100|12345", result_no_exclude)
-        self.assertIn("XYZ|1|100|67890", result_no_exclude)
-        self.assertIn("DEF|1|100|54321", result_no_exclude)
-        self.assertEqual(lines_count1, 3)
-        self.assertEqual(invalid_match_count1, 0)
-
-if __name__ == '__main__':
-    unittest.main()
